% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{clean_date}
\alias{clean_date}
\title{Clean Python datetime columns to R POSIXct}
\usage{
clean_date(column)
}
\arguments{
\item{column}{A vector (usually a column from a dataframe) containing
Python datetime objects or ISO-formatted character strings.}
}
\value{
A character vector of ISO date strings.
}
\description{
This function cleans a column of Python datetime objects (or strings)
returned via \code{reticulate}, converting them to ISO format strings for R.
}
\examples{
# Example with mixed data types:
clean_date(c("2025-04-23 00:00:00", ""))
}
