% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{install_pacifico_environment}
\alias{install_pacifico_environment}
\title{Install Python environment and 'pacifico' package}
\usage{
install_pacifico_environment()
}
\description{
This function installs Miniconda (if not available) and installs the 'pacifico'
Python package into the 'r-reticulate' environment.
}
