% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_pacifico_environment}
\alias{setup_pacifico_environment}
\title{Check Python environment and 'pacifico' availability}
\usage{
setup_pacifico_environment()
}
\value{
A list with the imported 'pacifico' and 'datetime' modules if available.
}
\description{
Verifies if the Python environment and 'pacifico' package are available.
Warns the user if setup is incomplete.
}
\keyword{internal}
