% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{to_py_date}
\alias{to_py_date}
\title{Convert R Date or POSIXct to Python datetime.date}
\usage{
to_py_date(x)
}
\arguments{
\item{x}{An R \code{Date}, \code{POSIXct}, or \code{POSIXlt} object.}
}
\value{
A Python \code{datetime.date} object.
}
\description{
This function converts an R \code{Date}, \code{POSIXct}, or \code{POSIXlt} object into a Python \code{datetime.date} object.
}
\examples{
to_py_date(as.Date("2025-04-23"))
}
