% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_pacifico_environment}
\alias{setup_pacifico_environment}
\title{Setup Pacifico Python Environment Safely}
\usage{
setup_pacifico_environment()
}
\value{
A list with the imported \code{pacifico} and \code{datetime} Python modules.
}
\description{
This function checks for Miniconda, creates the \code{r-reticulate} environment if needed,
installs the required Python package \code{pacifico}, and imports the necessary modules.
}
